fix f64 problems

+++ impy/random.py
@@ -184,7 +184,7 @@ class ImageGenerator:
     ) -> ImgArray:
         size, name, axes = _normalize_like(size, name, axes, like)
 
-        arr = self._rng.normal(loc=loc, scale=scale, size=size)
+        arr = self._rng.normal(loc=loc, scale=scale, size=size).astype(np.float32)
         if np.isscalar(arr):
             return arr
         return asarray(xp.asnumpy(arr), axes=axes, name=name)
